Code refactoring, adding new packages and removing useless code.
