--- api/src/main/java/com/guztaver/nexz/restservice/Greeting.java
@@ -1,2 +1,0 @@
-package com.guztaver.nexz.restservice;
-public record Greeting(long id, String content) { }
