+++ api/src/main/java/com/guztaver/nexz/restservice/repositories/CharacterRepository.java
@@ -1,6 +1,6 @@
-package com.guztaver.nexz.sqlhandler;
+package com.guztaver.nexz.restservice.repositories;
 
-import com.guztaver.nexz.restservice.Character;
+import com.guztaver.nexz.restservice.entities.Character;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CharacterRepository extends CrudRepository<Character, Integer> {}
